requirements.txt
Django==1.7.3
django-allauth==0.19.1
django-appconf==1.0.1
django-braces==1.4.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-extensions==1.5.0
django-floppyforms==1.3.0
django-geojson==2.7.0
django-geoposition==0.2.2
django-imagekit==3.2.6
django-leaflet==0.15.2
django-registration==1.0
docutils==0.12
MySQL-python==1.2.5
oauthlib==0.7.2
pilkit==1.1.12
Pillow==2.7.0
pylibmc==1.4.1
python-openid==2.2.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.1.14
